# improve performance of print(DevNull, x)