We add a specialization of `unsafe_write` for `DevNullStream` and modify `write` to avoid a `sgt -1` check which throws an exception on failure. The second change is safe because the len field of a string can be assumed to be a nonnegative integer, and indeed is already assumed to be nonnegative in other places.

Before:

```
julia> @benchmark print(DevNull, "foo")
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     5.974 ns (0.00% GC)
  median time:      5.982 ns (0.00% GC)
  mean time:        6.090 ns (0.00% GC)
  maximum time:     20.814 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

After:

```julia
julia> @benchmark print(DevNull, "foo")
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     1.925 ns (0.00% GC)
  median time:      1.933 ns (0.00% GC)
  mean time:        1.978 ns (0.00% GC)
  maximum time:     13.797 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1000
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

in particular,

```julia
julia> @code_llvm print(DevNull, "foo")

define void @julia_print_68274(i8**) #0 !dbg !5 {
top:
  ret void
}
```